+++ engine/src/dpx_tt_gsub.rs
@@ -394,7 +394,7 @@ unsafe extern "C" fn clt_read_record_list(
 unsafe extern "C" fn clt_release_record_list(mut list: *mut clt_record_list) {
     if !list.is_null() {
         (*list).record = mfree((*list).record as *mut libc::c_void) as *mut clt_record;
-        (*list).count = 0i32 as u16
+        (*list).count = 0_u16
     };
 }
 unsafe extern "C" fn clt_read_number_list(
@@ -432,7 +432,7 @@ unsafe extern "C" fn clt_read_number_list(
 unsafe extern "C" fn clt_release_number_list(mut list: *mut clt_number_list) {
     if !list.is_null() {
         (*list).value = mfree((*list).value as *mut libc::c_void) as *mut u16;
-        (*list).count = 0i32 as u16
+        (*list).count = 0_u16
     };
 }
 unsafe extern "C" fn clt_read_script_table(
@@ -611,7 +611,7 @@ unsafe extern "C" fn clt_release_coverage(mut cov: *mut clt_coverage) {
             }
         }
     }
-    (*cov).count = 0i32 as u16;
+    (*cov).count = 0_u16;
 }
 unsafe extern "C" fn clt_lookup_coverage(mut cov: *mut clt_coverage, mut gid: u16) -> i32 {
     let mut i: i32 = 0;
@@ -680,7 +680,7 @@ unsafe extern "C" fn otl_gsub_read_single(
         );
     }
     offset = ttstub_input_seek((*sfont).handle, 0i32 as ssize_t, 1i32) as u32;
-    (*subtab).LookupType = 1i32 as u16;
+    (*subtab).LookupType = 1_u16;
     (*subtab).SubstFormat = tt_get_unsigned_pair((*sfont).handle);
     len = 2i32;
     if (*subtab).SubstFormat as i32 == 1i32 {
@@ -712,7 +712,7 @@ unsafe extern "C" fn otl_gsub_read_single(
             (*data_0).Substitute = new(((*data_0).GlyphCount as u32 as u64)
                 .wrapping_mul(::std::mem::size_of::<GlyphID>() as u64)
                 as u32) as *mut GlyphID;
-            count = 0i32 as u16;
+            count = 0_u16;
             while (count as i32) < (*data_0).GlyphCount as i32 {
                 *(*data_0).Substitute.offset(count as isize) =
                     tt_get_unsigned_pair((*sfont).handle);
@@ -758,7 +758,7 @@ unsafe extern "C" fn otl_gsub_read_alternate(
         );
     }
     offset = ttstub_input_seek((*sfont).handle, 0i32 as ssize_t, 1i32) as u32;
-    (*subtab).LookupType = 3i32 as u16;
+    (*subtab).LookupType = 3_u16;
     (*subtab).SubstFormat = tt_get_unsigned_pair((*sfont).handle);
     if (*subtab).SubstFormat as i32 != 1i32 {
         dpx_warning(
@@ -777,15 +777,15 @@ unsafe extern "C" fn otl_gsub_read_alternate(
     (*data).AlternateSetCount = altset_offsets.count;
     if (*data).AlternateSetCount as i32 == 0i32 {
         (*data).AlternateSet = 0 as *mut otl_gsub_altset;
-        (*data).coverage.count = 0i32 as u16;
-        (*data).coverage.format = 0i32 as u16;
+        (*data).coverage.count = 0_u16;
+        (*data).coverage.format = 0_u16;
         (*data).coverage.list = 0 as *mut GlyphID;
         return len;
     }
     (*data).AlternateSet = new(((*data).AlternateSetCount as u32 as u64)
         .wrapping_mul(::std::mem::size_of::<otl_gsub_altset>() as u64)
         as u32) as *mut otl_gsub_altset;
-    i = 0i32 as u16;
+    i = 0_u16;
     while (i as i32) < (*data).AlternateSetCount as i32 {
         let mut altset: *mut otl_gsub_altset = 0 as *mut otl_gsub_altset;
         let mut altset_offset: u32 = 0;
@@ -801,7 +801,7 @@ unsafe extern "C" fn otl_gsub_read_alternate(
             (*altset).Alternate = new(((*altset).GlyphCount as u32 as u64)
                 .wrapping_mul(::std::mem::size_of::<GlyphID>() as u64)
                 as u32) as *mut GlyphID;
-            j = 0i32 as u16;
+            j = 0_u16;
             while (j as i32) < (*altset).GlyphCount as i32 {
                 *(*altset).Alternate.offset(j as isize) = tt_get_unsigned_pair((*sfont).handle);
                 len += 2i32;
@@ -846,7 +846,7 @@ unsafe extern "C" fn otl_gsub_read_ligature(
         );
     }
     offset = ttstub_input_seek((*sfont).handle, 0i32 as ssize_t, 1i32) as u32;
-    (*subtab).LookupType = 4i32 as u16;
+    (*subtab).LookupType = 4_u16;
     (*subtab).SubstFormat = tt_get_unsigned_pair((*sfont).handle);
     if (*subtab).SubstFormat as i32 != 1i32 {
         dpx_warning(
@@ -865,15 +865,15 @@ unsafe extern "C" fn otl_gsub_read_ligature(
     (*data).LigSetCount = ligset_offsets.count;
     if (*data).LigSetCount as i32 == 0i32 {
         (*data).LigatureSet = 0 as *mut otl_gsub_ligset;
-        (*data).coverage.count = 0i32 as u16;
-        (*data).coverage.format = 0i32 as u16;
+        (*data).coverage.count = 0_u16;
+        (*data).coverage.format = 0_u16;
         (*data).coverage.list = 0 as *mut GlyphID;
         return len;
     }
     (*data).LigatureSet = new(((*data).LigSetCount as u32 as u64)
         .wrapping_mul(::std::mem::size_of::<otl_gsub_ligset>() as u64)
         as u32) as *mut otl_gsub_ligset;
-    i = 0i32 as u16;
+    i = 0_u16;
     while (i as i32) < (*data).LigSetCount as i32 {
         let mut ligset_tab: clt_number_list = clt_number_list {
             count: 0,
@@ -894,7 +894,7 @@ unsafe extern "C" fn otl_gsub_read_ligature(
             (*ligset).Ligature = new((ligset_tab.count as u32 as u64)
                 .wrapping_mul(::std::mem::size_of::<otl_gsub_ligtab>() as u64)
                 as u32) as *mut otl_gsub_ligtab;
-            j = 0i32 as u16;
+            j = 0_u16;
             while (j as i32) < ligset_tab.count as i32 {
                 ttstub_input_seek(
                     (*sfont).handle,
@@ -916,7 +916,7 @@ unsafe extern "C" fn otl_gsub_read_ligature(
                         - 1i32) as u32 as u64)
                         .wrapping_mul(::std::mem::size_of::<GlyphID>() as u64)
                         as u32) as *mut GlyphID;
-                    count = 0i32 as u16;
+                    count = 0_u16;
                     while (count as i32)
                         < (*(*ligset).Ligature.offset(j as isize)).CompCount as i32 - 1i32
                     {
@@ -977,11 +977,11 @@ unsafe extern "C" fn otl_gsub_release_ligature(mut subtab: *mut otl_gsub_subtab)
         let mut j: u16 = 0;
         data = (*subtab).table.ligature1;
         if !data.is_null() && !(*data).LigatureSet.is_null() {
-            i = 0i32 as u16;
+            i = 0_u16;
             while (i as i32) < (*data).LigSetCount as i32 {
                 let mut ligset: *mut otl_gsub_ligset = 0 as *mut otl_gsub_ligset;
                 ligset = &mut *(*data).LigatureSet.offset(i as isize) as *mut otl_gsub_ligset;
-                j = 0i32 as u16;
+                j = 0_u16;
                 while (j as i32) < (*ligset).LigatureCount as i32 {
                     let ref mut fresh2 = (*(*ligset).Ligature.offset(j as isize)).Component;
                     *fresh2 = mfree(
@@ -1007,7 +1007,7 @@ unsafe extern "C" fn otl_gsub_release_alternate(mut subtab: *mut otl_gsub_subtab
         let mut i: u16 = 0;
         data = (*subtab).table.alternate1;
         if !data.is_null() && !(*data).AlternateSet.is_null() {
-            i = 0i32 as u16;
+            i = 0_u16;
             while (i as i32) < (*data).AlternateSetCount as i32 {
                 let mut altset: *mut otl_gsub_altset = 0 as *mut otl_gsub_altset;
                 altset = &mut *(*data).AlternateSet.offset(i as isize) as *mut otl_gsub_altset;
@@ -1057,7 +1057,7 @@ unsafe extern "C" fn otl_gsub_read_feat(mut gsub: *mut otl_gsub_tab, mut sfont: 
         LookupList: 0,
     };
     let mut subtab: *mut otl_gsub_subtab = 0 as *mut otl_gsub_subtab;
-    let mut num_subtabs: u16 = 0i32 as u16;
+    let mut num_subtabs: u16 = 0_u16;
     let mut feat_bits: [u8; 8192] = [0; 8192];
     let mut feature_list: clt_record_list = clt_record_list {
         count: 0,
@@ -1597,7 +1597,7 @@ unsafe extern "C" fn glyph_seq_cmp(
     if n_glyphs0 as i32 != n_glyphs1 as i32 {
         return n_glyphs0 as i32 - n_glyphs1 as i32;
     }
-    i = 0i32 as u16;
+    i = 0_u16;
     while (i as i32) < n_glyphs0 as i32 {
         if *glyph_seq0.offset(i as isize) as i32 != *glyph_seq1.offset(i as isize) as i32 {
             return *glyph_seq0.offset(i as isize) as i32 - *glyph_seq1.offset(i as isize) as i32;
@@ -1633,7 +1633,7 @@ unsafe extern "C" fn otl_gsub_apply_ligature(
             let mut ligset: *mut otl_gsub_ligset = 0 as *mut otl_gsub_ligset;
             let mut j: u16 = 0;
             ligset = &mut *(*data).LigatureSet.offset(idx as isize) as *mut otl_gsub_ligset;
-            j = 0i32 as u16;
+            j = 0_u16;
             while (j as i32) < (*ligset).LigatureCount as i32 {
                 if glyph_seq_cmp(
                     &mut *gid_in.offset(1),
@@ -1756,11 +1756,11 @@ unsafe extern "C" fn scan_otl_tag(
         return -1i32;
     }
     memset(script as *mut libc::c_void, ' ' as i32, 4i32 as u64);
-    *script.offset(4) = 0i32 as i8;
+    *script.offset(4) = 0_i8;
     memset(language as *mut libc::c_void, ' ' as i32, 4i32 as u64);
-    *language.offset(4) = 0i32 as i8;
+    *language.offset(4) = 0_i8;
     memset(feature as *mut libc::c_void, ' ' as i32, 4i32 as u64);
-    *feature.offset(4) = 0i32 as i8;
+    *feature.offset(4) = 0_i8;
     /* First parse otl_tags variable */
     p = otl_tags;
     period = strchr(p, '.' as i32);
